requirements.txt
flask
flask-cors
pymongo
propelauth_fastapi
fastapi
openai==0.28
python-dotenv